requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
debtcollector==2.0.1
Django==3.0.5
django-cors-headers==3.2.1
django-model-utils==4.0.0
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
iso8601==0.1.12
keystoneauth1==4.0.0
msgpack==1.0.0
mysqlclient==1.4.6
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.7.0
oslo.config==8.0.2
oslo.i18n==4.0.1
oslo.serialization==3.1.1
oslo.utils==4.1.1
pbr==5.4.5
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dotenv==0.12.0
python-http-client==3.2.7
python-keystoneclient==4.0.0
python-swiftclient==3.9.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
rfc3986==1.4.0
sendgrid==6.2.1
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
urllib3==1.25.8
wrapt==1.12.1
